initial logger refactor to PSR-3
