+++ Core/Interfaces/Logger.php
@@ -1,41 +1,19 @@
 <?php
 namespace Core\Interfaces;
+use Psr\Log\LoggerInterface;
 /** 
 *	An interface defining a Logger
 *
 *	@author Jason Savell <jsavell@library.tamu.edu>
 */
 
-interface Logger {
+interface Logger extends LoggerInterface {
 	/**
-	*	Log an informational message about the state of the application
-	*
-	*	@param string $message The info message
-	*	@return void
-	*/
-	public function info($message);
-
-	/**
-	*	Log a message that may be helpful when debugging the application
-	*
-	*	@param string $message The debug message
-	*	@return void
-	*/
-	public function debug($message);
-
-	/**
+	* 	@deprecated Use warning(...) instead
 	*	Log a message that warns that something may be wrong with the application
 	*
 	*	@param string $message The warning message
 	*	@return void
 	*/
 	public function warn($message);
-
-	/**
-	*	Log a message that notifies that the application is/was in an error state.
-	*
-	*	@param string $message The error message
-	*	@return void
-	*/
-	public function error($message);
 }
